refactor: library split
