+++ deployer/src/deployment/mod.rs
@@ -1,13 +1,11 @@
 pub mod deploy_layer;
-mod provisioner_factory;
+pub mod provisioner_factory;
 mod queue;
 mod run;
-mod runtime_logger;
+pub mod runtime_logger;
 
-pub use provisioner_factory::AbstractProvisionerFactory;
 pub use queue::Queued;
 pub use run::Built;
-pub use runtime_logger::RuntimeLoggerFactory;
 use tracing::instrument;
 
 use crate::persistence::{SecretRecorder, State};
